Moving flask to backend
- moved Flask to a new location: `backend`
- code adjustments
- added and updated targets in Makefile
