+++ Makefile
@@ -1,42 +1,53 @@
 VENV = venv
+BACKEND = backend
 DATA_COLLECTOR = data_collector
-VENV_DATA_COLLECTOR = $(DATA_COLLECTOR)/venv
-PYTHON_VERSION = 3.9
-SYSTEM_PYTHON = $(shell which python$(PYTHON_VERSION))
-PYTHON = $(VENV)/bin/python3
-PYTHON_DATA_COLLECTOR = $(VENV_DATA_COLLECTOR)/bin/python3
-PIP = $(VENV)/bin/pip
-PIP_DATA_COLLECTOR = $(VENV_DATA_COLLECTOR)/bin/pip
 FRONTEND = frontend
 
+VENV_BACKEND = $(BACKEND)/$(VENV)
+VENV_DATA_COLLECTOR = $(DATA_COLLECTOR)/$(VENV)
+
+PYTHON_VERSION = 3.9
+SYSTEM_PYTHON = $(shell which python$(PYTHON_VERSION))
+
+VENV_ACTIVATE_BACKEND = . $(VENV_BACKEND)/bin/activate
+VENV_ACTIVATE_DATA_COLLECTOR = . $(VENV_DATA_COLLECTOR)/bin/activate
+
+PYTHON_BACKEND = $(VENV_BACKEND)/bin/python3
+PYTHON_DATA_COLLECTOR = $(VENV_DATA_COLLECTOR)/bin/python3
+
+ENV_VARS = set -o allexport; . $(BACKEND)/.env; set +o allexport
+
+PIP_BACKEND = $(VENV_BACKEND)/bin/pip
+PIP_DATA_COLLECTOR = $(VENV_DATA_COLLECTOR)/bin/pip
+
 venv:
-	test -d $(VENV) || $(SYSTEM_PYTHON) -m venv $(VENV)
+	test -d $(VENV_BACKEND) || $(SYSTEM_PYTHON) -m venv $(VENV_BACKEND)
 	test -d $(VENV_DATA_COLLECTOR) || $(SYSTEM_PYTHON) -m venv $(VENV_DATA_COLLECTOR)
 
 install: venv
-	. $(VENV)/bin/activate && $(PIP) install -r requirements.txt
-	. $(VENV_DATA_COLLECTOR)/bin/activate && $(PIP_DATA_COLLECTOR) install -r $(DATA_COLLECTOR)/requirements.txt
+	$(VENV_ACTIVATE_BACKEND) && $(PIP_BACKEND) install -r $(BACKEND)/requirements.txt
+	$(VENV_ACTIVATE_DATA_COLLECTOR) && $(PIP_DATA_COLLECTOR) install -r $(DATA_COLLECTOR)/requirements.txt
 
 install-dev: install
-	. $(VENV)/bin/activate $(PIP) install -r requirements_dev.txt
-	. $(VENV_DATA_COLLECTOR)/bin/activate && $(PIP_DATA_COLLECTOR) install -r $(DATA_COLLECTOR)/requirements_dev.txt
+	$(VENV_ACTIVATE_BACKEND) && $(PIP_BACKEND) install -r $(BACKEND)/requirements_dev.txt
+	$(VENV_ACTIVATE_DATA_COLLECTOR) && $(PIP_DATA_COLLECTOR) install -r $(DATA_COLLECTOR)/requirements_dev.txt
 
 install-ui:
 	npm install --prefix $(FRONTEND)
 
 lint:
-	. $(VENV)/bin/activate && $(VENV)/bin/pylint config.py wsgi.py application/
-	. $(VENV_DATA_COLLECTOR)/bin/activate && $(VENV_DATA_COLLECTOR)/bin/pylint $(DATA_COLLECTOR)/data_collector.py
+	$(VENV_ACTIVATE_BACKEND) && $(VENV_BACKEND)/bin/pylint --rcfile=$(BACKEND)/.pylintrc $(BACKEND)/config.py $(BACKEND)/wsgi.py $(BACKEND)/application/
+	$(VENV_ACTIVATE_DATA_COLLECTOR) && $(VENV_DATA_COLLECTOR)/bin/pylint --rcfile=$(DATA_COLLECTOR)/.pylintrc $(DATA_COLLECTOR)/data_collector.py
 
 lint-ui:
 	npm run --prefix $(FRONTEND) lint
 
 flake8:
-	. $(VENV)/bin/activate && $(VENV)/bin/flake8 config.py wsgi.py application/
-	. $(VENV_DATA_COLLECTOR)/bin/activate && $(VENV_DATA_COLLECTOR)/bin/flake8 $(DATA_COLLECTOR)/data_collector.py
+	$(VENV_ACTIVATE_BACKEND) && $(VENV_BACKEND)/bin/flake8 --config=$(BACKEND)/.flake8 $(BACKEND)/config.py $(BACKEND)/wsgi.py $(BACKEND)/application/
+	$(VENV_ACTIVATE_DATA_COLLECTOR) && $(VENV_DATA_COLLECTOR)/bin/flake8 --config=$(DATA_COLLECTOR)/.flake8 $(DATA_COLLECTOR)/data_collector.py
 
 testing:
-	. $(VENV)/bin/activate && pytest --cov=application tests/
+	$(VENV_ACTIVATE_BACKEND) && $(VENV_BACKEND)/bin/pytest --cov=$(BACKEND)/application $(BACKEND)/tests/
 
 testing-ui:
 	npm run --prefix $(FRONTEND) test:unit
@@ -44,43 +55,41 @@ testing-ui:
 checking: lint flake8 testing lint-ui testing-ui
 
 create-env:
-ifeq ($(shell test -s .env && echo -n 0), 0)
-	@echo 'Nothing to be done for create-env - .env file exists.'
+ifeq ($(shell test -s $(BACKEND)/.env && echo -n 0), 0)
+	@echo 'Skipping this step - backend/.env file exists.'
 else
-	cp .env.example .env
-	. $(VENV)/bin/activate && $(PYTHON) -c 'import os; print("SECRET_KEY={}".format(os.urandom(24)))' >> .env
+	cp $(BACKEND)/.env.example $(BACKEND)/.env
+	$(VENV_ACTIVATE_BACKEND) && $(PYTHON_BACKEND) -c 'import os; print("SECRET_KEY={}".format(os.urandom(24)))' >> $(BACKEND)/.env
 endif
 
 create-config:
-ifeq ($(shell test -s data-usage-monitor.ini && echo -n 0), 0)
-	@echo 'Nothing to be done for create-config - data-usage-monitor.ini file exists.'
+ifeq ($(shell test -s $(BACKEND)/data-usage-monitor.ini && echo -n 0), 0)
+	@echo 'Skipping this step - backend/data-usage-monitor.ini file exists.'
 else
-	cp data-usage-monitor.ini.example data-usage-monitor.ini
+	cp data-usage-monitor.ini.example $(BACKEND)/data-usage-monitor.ini
 endif
-
-copy-config:
 ifeq ($(shell test -s $(DATA_COLLECTOR)/data-usage-monitor.ini && echo -n 0), 0)
-	@echo 'Nothing to be done for copy-config - data_collector/data-usage-monitor.ini file exists.'
+	@echo 'Skipping this step - data_collector/data-usage-monitor.ini file exists.'
 else
-	cp data-usage-monitor.ini $(DATA_COLLECTOR)/data-usage-monitor.ini
+	cp data-usage-monitor.ini.example $(DATA_COLLECTOR)/data-usage-monitor.ini
 endif
 
 prepare: install install-ui create-env create-config
 
 upgrade-db:
-	. $(VENV)/bin/activate && flask db upgrade
+	$(VENV_ACTIVATE_BACKEND) && $(ENV_VARS) && $(VENV_BACKEND)/bin/flask db upgrade --directory $(BACKEND)/migrations
 
 downgrade-db:
-	. $(VENV)/bin/activate && flask db downgrade base
+	$(VENV_ACTIVATE_BACKEND) && $(ENV_VARS) && $(VENV_BACKEND)/bin/flask db downgrade base --directory $(BACKEND)/migrations
 
 run-flask:
-	. $(VENV)/bin/activate && $(PYTHON) wsgi.py
+	$(VENV_ACTIVATE_BACKEND) && $(ENV_VARS) && $(VENV_BACKEND)/bin/flask run
 
 run-uwsgi:
-	. $(VENV)/bin/activate && uwsgi --ini data-usage-monitor.ini
+	$(VENV_ACTIVATE_BACKEND) && $(VENV_BACKEND)/bin/uwsgi --ini $(BACKEND)/data-usage-monitor.ini
 
 run-data-collector:
-	. $(VENV_DATA_COLLECTOR)/bin/activate && $(PYTHON_DATA_COLLECTOR) $(DATA_COLLECTOR)/data_collector.py
+	$(VENV_ACTIVATE_DATA_COLLECTOR) && $(PYTHON_DATA_COLLECTOR) $(DATA_COLLECTOR)/data_collector.py
 
 run-ui:
 	export VUE_APP_USE_MOCKED_VALUES=false && npm run --prefix $(FRONTEND) serve
@@ -92,7 +101,12 @@ clean:
 	rm -rf __pycache__
 	rm -rf $(VENV)
 	rm -rf $(DATA_COLLECTOR)/__pycache__
+	rm -rf $(DATA_COLLECTOR)/data-usage-monitor.ini
 	rm -rf $(VENV_DATA_COLLECTOR)
+	rm -rf $(BACKEND)/__pycache__
+	rm -rf $(BACKEND)/data-usage-monitor.ini
+	rm -rf $(BACKEND)/.env
+	rm -rf $(VENV_BACKEND)
 	rm -rf $(FRONTEND)/node_modules
 	
 .PHONY: venv install install-dev install-ui lint lint-ui flake8 testing testing-ui checking create-env create-config copy-config prepare upgrade-db run-app run-data-collector run-ui run-ui-mock clean
