+++ backend/requirements.txt
@@ -1,8 +1,4 @@
-bs4==0.0.1
 Flask==2.1.1
-lxml==4.8.0
-requests==2.27.1
-schedule==1.1.0
 python-dotenv==0.20.0
 psycopg2==2.9.3
 SQLAlchemy==1.4.35
