+++ backend/tests/test_blueprints.py
@@ -1,5 +1,5 @@
 """
-Set of tests for application.
+Set of tests for backend.application.
 """
 from decimal import Decimal
 import json
@@ -10,7 +10,7 @@ def test_index(client, mocker):
     mocked_usage_stamp.current_month_download = Decimal(10.500)
     mocked_usage_stamp.current_month_upload = Decimal(4.0)
     mocked_usage_stamp.time_stamp = 'Sun, 29 May 2022 14:22:34 GMT'
-    mock_get_newest_by_date = mocker.patch('application.models.UsageStamp.get_newest_by_date',
+    mock_get_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_newest_by_date',
                                            return_value=mocked_usage_stamp)
     response = client.get('/')
     mock_get_newest_by_date.assert_called_once()
@@ -19,7 +19,7 @@ def test_index(client, mocker):
 
 
 def test_index_without_usage_stamps(client, mocker):
-    mock_get_newest_by_date = mocker.patch('application.models.UsageStamp.get_newest_by_date',
+    mock_get_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_newest_by_date',
                                            return_value=None)
     response = client.get('/')
     mock_get_newest_by_date.assert_called_once()
@@ -38,7 +38,7 @@ def test_get_usage_stamp(client, mocker):
                             'current_month_upload': '4.0',
                             'time_stamp': 'Sun, 29 May 2022 14:22:34 GMT'}
 
-    mock_get_newest_by_date = mocker.patch('application.models.UsageStamp.get_newest_by_date',
+    mock_get_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_newest_by_date',
                                            return_value=mocked_usage_stamp)
     response = client.get('/api/usage-stamp')
     mock_get_newest_by_date.assert_called_once()
@@ -48,7 +48,7 @@ def test_get_usage_stamp(client, mocker):
 
 def test_get_usage_stamp_without_usage_stamps(client, mocker):
 
-    mock_get_newest_by_date = mocker.patch('application.models.UsageStamp.get_newest_by_date',
+    mock_get_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_newest_by_date',
                                            return_value=None)
     response = client.get('/api/usage-stamp')
     mock_get_newest_by_date.assert_called_once()
@@ -60,8 +60,8 @@ def test_put_usage_stamp(client, mocker):
     mocked_usage_stamp = {'current_month_download': '10.5',
                           'current_month_upload': '4.0',
                           'time_stamp': 'Sun, 29 May 2022 14:22:34 GMT'}
-    mock_db_session_add = mocker.patch('application.extensions.db.session.add')
-    mock_db_session_commit = mocker.patch('application.extensions.db.session.commit')
+    mock_db_session_add = mocker.patch('backend.application.extensions.db.session.add')
+    mock_db_session_commit = mocker.patch('backend.application.extensions.db.session.commit')
     
     response = client.put('/api/usage-stamp', query_string=mocked_usage_stamp)
     mock_db_session_add.assert_called_once()
@@ -74,8 +74,8 @@ def test_put_usage_stamp_bad_request(client, mocker):
     mocked_usage_stamp = {'current_month_download': None,
                           'current_month_upload': '4.0',
                           'time_stamp': 'Sun, 29 May 2022 14:22:34 GMT'}
-    mock_db_session_add = mocker.patch('application.extensions.db.session.add')
-    mock_db_session_commit = mocker.patch('application.extensions.db.session.commit')
+    mock_db_session_add = mocker.patch('backend.application.extensions.db.session.add')
+    mock_db_session_commit = mocker.patch('backend.application.extensions.db.session.commit')
     
     response = client.put('/api/usage-stamp', query_string=mocked_usage_stamp)
     mock_db_session_add.assert_not_called()
@@ -92,7 +92,7 @@ def test_get_usage_stamps(client, mocker):
                             'current_month_upload': '4.0',
                             'time_stamp': 'Sun, 29 May 2022 14:26:34 GMT'}
 
-    mock_get_multiple_newest_by_date = mocker.patch('application.models.UsageStamp.get_multiple_newest_by_date',
+    mock_get_multiple_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_multiple_newest_by_date',
                                            return_value=[mocked_usage_stamp_1, mocked_usage_stamp_2])
     response = client.get('/api/usage-stamps')
     mock_get_multiple_newest_by_date.assert_called_once()
@@ -103,7 +103,7 @@ def test_get_usage_stamps(client, mocker):
 
 
 def test_get_usage_stamps_without_usage_stamps(client, mocker):
-    mock_get_multiple_newest_by_date = mocker.patch('application.models.UsageStamp.get_multiple_newest_by_date',
+    mock_get_multiple_newest_by_date = mocker.patch('backend.application.models.UsageStamp.get_multiple_newest_by_date',
                                            return_value=[])
     response = client.get('/api/usage-stamps')
     mock_get_multiple_newest_by_date.assert_called_once()
